+++ web/src/core/BeamUtils.ts
@@ -1,5 +1,7 @@
 import type { TokenResponse } from '@/__generated__/schemas';
 import { TokenStorage } from '@/platform/types/TokenStorage';
+import type { RefreshableServiceMap } from '@/core/types';
+import { BeamJsonUtils } from '@/utils/BeamJsonUtils';
 
 /** Saves the access token, refresh token, and expiration time from a token response to the token storage. */
 export async function saveToken(
@@ -11,4 +13,22 @@ export async function saveToken(
   await tokenStorage.setExpiresIn(
     Date.now() + Number(tokenResponse.expires_in),
   );
+}
+
+/** Parses a raw socket message string into a structured object based on the expected `RefreshableServiceMap` type. */
+export function parseSocketMessage<K extends keyof RefreshableServiceMap>(
+  rawMessage: string,
+): RefreshableServiceMap[K] {
+  const parsed = JSON.parse(rawMessage, BeamJsonUtils.reviver);
+
+  const { scopes, delay, ...rest } = parsed;
+
+  // The unknown field will be the remaining key in `rest`
+  const [unknownKey] = Object.keys(rest);
+  const data = rest[unknownKey] as RefreshableServiceMap[K]['data'];
+
+  return {
+    data,
+    ...parsed,
+  };
 }